+++ tsfresh/feature_extraction/feature_calculators.py
@@ -665,27 +665,41 @@ def sum_of_reoccurring_values(x):
     :return: the value of this feature
     :return type: float
     """
-    unique, counts = np.unique(x, return_counts=True)
-    counts[counts < 2] = 0
-    counts[counts > 1] = 1
-    return np.sum(counts * unique)
+    # @modified 20201230 - Branch #3900: v0.5.1
+    # Revert to the original sum_of_reoccurring_values v0.4.0 method which was
+    # changed and the new feature called sum_of_reoccurring_data_points was
+    # added which results in the same value as the original v0.4.0
+    # sum_of_reoccurring_values method. The new sum_of_reoccurring_values method
+    # introduced results in different results as per:
+    # NOT in baseline   :: [['value__sum_of_reoccurring_values', '49922.0']]
+    # NOT in calculated :: [['value__sum_of_reoccurring_values', '109822.0']]
+    # unique, counts = np.unique(x, return_counts=True)
+    # counts[counts < 2] = 0
+    # counts[counts > 1] = 1
+    # return np.sum(counts * unique)
+    x = pd.Series(x)
+    value_counts = x.value_counts()
+    doubled_values = value_counts[value_counts > 1]
+    return sum(doubled_values.index * doubled_values)
 
 
-@set_property("fctype", "aggregate")
-@not_apply_to_raw_numbers
-def sum_of_reoccurring_data_points(x):
-    """
-    Returns the sum of all data points, that are present in the time series
-    more than once.
-
-    :param x: the time series to calculate the feature of
-    :type x: pandas.Series
-    :return: the value of this feature
-    :return type: float
-    """
-    unique, counts = np.unique(x, return_counts=True)
-    counts[counts < 2] = 0
-    return np.sum(counts * unique)
+# @modified 20201230 - Branch #3900: v0.5.1
+# Disable sum_of_reoccurring_data_points feature
+# @set_property("fctype", "aggregate")
+# @not_apply_to_raw_numbers
+# def sum_of_reoccurring_data_points(x):
+#    """
+#    Returns the sum of all data points, that are present in the time series
+#    more than once.
+#
+#    :param x: the time series to calculate the feature of
+#    :type x: pandas.Series
+#    :return: the value of this feature
+#    :return type: float
+#    """
+#    unique, counts = np.unique(x, return_counts=True)
+#    counts[counts < 2] = 0
+#    return np.sum(counts * unique)
 
 
 @set_property("fctype", "aggregate")
@@ -1288,124 +1302,127 @@ def approximate_entropy(x, m, r):
 
     return np.abs(_phi(m) - _phi(m + 1))
 
-def _estimate_friedrich_coefficients(x, m, r):
-    """
-    Coefficients of polynomial :math:`h(x)`, which has been fitted to 
-    the deterministic dynamics of Langevin model 
-    .. math::
-        \dot{x}(t) = h(x(t)) + \mathcal{N}(0,R)
+# @modified 20201230 - Branch #3902: v0.6.1
+# Disabled _estimate_friedrich_coefficients introduced in v0.6.0
+# def _estimate_friedrich_coefficients(x, m, r):
+#     """
+#     Coefficients of polynomial :math:`h(x)`, which has been fitted to
+#     the deterministic dynamics of Langevin model
+#     .. math::
+#         \dot{x}(t) = h(x(t)) + \mathcal{N}(0,R)
+#
+#     As described by
+#
+#         Friedrich et al. (2000): Physics Letters A 271, p. 217-222
+#         *Extracting model equations from experimental data*
+#
+#     For short time-series this method is highly dependent on the parameters.
+#
+#     :param x: the time series to calculate the feature of
+#     :type x: pandas.Series
+#     :param m: order of polynom to fit for estimating fixed points of dynamics
+#     :type m: int
+#     :param r: number of quantils to use for averaging
+#     :type r: float
+#
+#     :return: coefficients of polynomial of deterministic dynamics
+#     :return type: ndarray
+#     """
+#     df = pd.DataFrame({'signal': x[:-1], 'delta': np.diff(x)})
+#     try:
+#         df['quantiles']=pd.qcut(df.signal, r)
+#         binned = True
+#     except ValueError:
+#         binned = False
+#         coeff = [np.NaN] * (m+1)
+#
+#     if binned:
+#         quantiles = df.groupby('quantiles')
+#
+#         result = pd.DataFrame({'x_mean': quantiles.signal.mean(),
+#                                'y_mean': quantiles.delta.mean()
+#         })
+#
+#         try:
+#             coeff = np.polyfit(result.x_mean, result.y_mean, deg=m)
+#         except (np.linalg.LinAlgError, ValueError):
+#             coeff = [np.NaN] * (m+1)
+#     return coeff
 
-    As described by
+# @modified 20201230 - Branch #3902: v0.6.1
+# Disabled friedrich_coefficients introduced in v0.6.0
+# @set_property("fctype", "apply")
+# def friedrich_coefficients(x, c, param):
+#     """
+#     Coefficients of polynomial :math:`h(x)`, which has been fitted to
+#     the deterministic dynamics of Langevin model
+#     .. math::
+#         \dot{x}(t) = h(x(t)) + \mathcal{N}(0,R)
+#
+#     as described by
+#
+#         Friedrich et al. (2000): Physics Letters A 271, p. 217-222
+#         *Extracting model equations from experimental data*
+#
+#
+#     For short time-series this method is highly dependent on the parameters.
+#
+#     :param x: the time series to calculate the feature of
+#     :type x: pandas.Series
+#     :param c: the time series name
+#     :type c: str
+#     :param param: contains dictionaries {"coeff": x} with x int and x >= 0
+#     :type param: list
+#     :return: the different feature values
+#     :return type: pandas.Series
+#     """
+#     coefficients = set([config["coeff"] for config in param])
+#     for coeff in coefficients:
+#         if coeff < 0:
+#             raise ValueError("Coefficients must be positive or zero.")
+#
+#     m = param[0]['m']
+#     r = param[0]['r']
+#
+#     coeff = _estimate_friedrich_coefficients(x, m, r)
+#
+#     name = lambda q: "{}__friedrich_coefficients__m_{}__r_{}__coeff_{}".format(c,m,r,q)
+#     return pd.Series(coeff, index=[name(q) for q in range(m,-1,-1)])
 
-        Friedrich et al. (2000): Physics Letters A 271, p. 217-222
-        *Extracting model equations from experimental data*
+# @modified 20201230 - Branch #3902: v0.6.1
+# Disabled max_langevin_fixed_point introduced in v0.6.0
 
-    For short time-series this method is highly dependent on the parameters.
-
-    :param x: the time series to calculate the feature of
-    :type x: pandas.Series
-    :param m: order of polynom to fit for estimating fixed points of dynamics
-    :type m: int
-    :param r: number of quantils to use for averaging
-    :type r: float
-
-    :return: coefficients of polynomial of deterministic dynamics
-    :return type: ndarray
-    """
-    df = pd.DataFrame({'signal': x[:-1], 'delta': np.diff(x)})
-    try:
-        df['quantiles'] = pd.qcut(df.signal, r)
-        binned = True
-    except ValueError:
-        binned = False
-        coeff = [np.NaN] * (m+1)
-
-    if binned:
-        quantiles = df.groupby('quantiles')
-        
-        result = pd.DataFrame({'x_mean': quantiles.signal.mean(),
-                               'y_mean': quantiles.delta.mean()
-        })
-
-        result.dropna(inplace=True)
-
-        try:
-            coeff = np.polyfit(result.x_mean, result.y_mean, deg=m)
-        except (np.linalg.LinAlgError, ValueError):
-            coeff = [np.NaN] * (m+1)
-    return coeff
-
-@set_property("fctype", "apply")
-def friedrich_coefficients(x, c, param):
-    """
-    Coefficients of polynomial :math:`h(x)`, which has been fitted to 
-    the deterministic dynamics of Langevin model 
-    .. math::
-        \dot{x}(t) = h(x(t)) + \mathcal{N}(0,R)
-
-    as described by
-
-        Friedrich et al. (2000): Physics Letters A 271, p. 217-222
-        *Extracting model equations from experimental data*
-
-
-    For short time-series this method is highly dependent on the parameters.
-
-    :param x: the time series to calculate the feature of
-    :type x: pandas.Series
-    :param c: the time series name
-    :type c: str
-    :param param: contains dictionaries {"coeff": x} with x int and x >= 0
-    :type param: list
-    :return: the different feature values
-    :return type: pandas.Series
-    """
-    coefficients = set([config["coeff"] for config in param])
-    for coeff in coefficients:
-        if coeff < 0:
-            raise ValueError("Coefficients must be positive or zero.")
-
-    m = param[0]['m']
-    r = param[0]['r']
-
-    coeff = _estimate_friedrich_coefficients(x, m, r)
-
-    name = lambda q: "{}__friedrich_coefficients__m_{}__r_{}__coeff_{}".format(c,m,r,q)
-    return pd.Series(coeff, index=[name(q) for q in range(m,-1,-1)])
-
-@set_property("fctype", "aggregate_with_parameters")
-def max_langevin_fixed_point(x, r, m):
-    """
-    Largest fixed point of dynamics  :math:argmax_x {h(x)=0}` estimated from polynomial :math:`h(x)`, 
-    which has been fitted to the deterministic dynamics of Langevin model
-    .. math::
-        \dot(x)(t) = h(x(t)) + R \mathcal(N)(0,1)
-
-    as described by
-
-        Friedrich et al. (2000): Physics Letters A 271, p. 217-222
-        *Extracting model equations from experimental data*
-
-    For short time-series this method is highly dependent on the parameters.
-
-    :param x: the time series to calculate the feature of
-    :type x: pandas.Series
-    :param m: order of polynom to fit for estimating fixed points of dynamics
-    :type m: int
-    :param r: number of quantils to use for averaging
-    :type r: float
-
-    :return: Largest fixed point of deterministic dynamics
-    :return type: float
-    """
-
-    coeff = _estimate_friedrich_coefficients(x, m, r)
-
-    try:
-        max_fixed_point = np.max(np.real(np.roots(coeff)))
-    except (np.linalg.LinAlgError, ValueError):
-        return np.nan
-    
-    return max_fixed_point
-
-
+# @set_property("fctype", "aggregate_with_parameters")
+# def max_langevin_fixed_point(x, r, m):
+#     """
+#     Largest fixed point of dynamics  :math:argmax_x {h(x)=0}` estimated from polynomial :math:`h(x)`,
+#     which has been fitted to the deterministic dynamics of Langevin model
+#     .. math::
+#         \dot(x)(t) = h(x(t)) + R \mathcal(N)(0,1)
+#
+#     as described by
+#
+#         Friedrich et al. (2000): Physics Letters A 271, p. 217-222
+#         *Extracting model equations from experimental data*
+#
+#     For short time-series this method is highly dependent on the parameters.
+#
+#     :param x: the time series to calculate the feature of
+#     :type x: pandas.Series
+#     :param m: order of polynom to fit for estimating fixed points of dynamics
+#     :type m: int
+#     :param r: number of quantils to use for averaging
+#     :type r: float
+#
+#     :return: Largest fixed point of deterministic dynamics
+#     :return type: float
+#     """
+#
+#     coeff = _estimate_friedrich_coefficients(x, m, r)
+#
+#     try:
+#         max_fixed_point = np.max(np.real(np.roots(coeff)))
+#     except (np.linalg.LinAlgError, ValueError):
+#         return np.nan
+#
+#     return max_fixed_point