+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 
 tsfresh uses `Semantic Versioning <http://semver.org/>`_
 
+Version 0.7.2
+=============
+
+- readded baseline unit tests
+
 Version 0.7.1
 =============
 
@@ -30,6 +35,12 @@ Version 0.7.0
     - an index with same name as id_column was breaking parallelization
     - friedrich_coefficients and max_langevin_fixed_point were occasionally stalling
 
+Version 0.6.1
+=============
+
+- Disabled new feature: estimation of largest fixed point of deterministic dynamics
+- Disabled _esitmate_friedrich_coefficients, friedrich_coefficients and max_langevin_fixed_point
+
 Version 0.6.0
 =============
 
@@ -38,6 +49,11 @@ Version 0.6.0
 - new notebook: demonstration how to use tsfresh in a pipeline with train and test datasets
 - remove no logging handler warning
 - fixed bug in the RelevantFeatureAugmenter regarding the evaluate_only_added_features parameters
+
+Version 0.5.1
+=============
+
+- Disabled new feature and feature renaming: sum_of_recurring_values, sum_of_recurring_data_points
 
 Version 0.5.0
 =============